added weights  and online network as data in library wrapper
