+++ bark_ml/library_wrappers/lib_fqf_iqn_qrdqn/model_wrapper/model_data/BUILD
@@ -1,0 +1,5 @@
+filegroup(
+    name = "weights",
+    srcs = glob(["*.pt", "*.pth"]),
+    visibility = ["//visibility:public"],
+)
